requirements.txt
numpy~=2.1.3
numba~=0.61.0
grpcio~=1.71.0
protobuf~=5.29.3
grpcio-reflection~=1.71.0
grpcio-tools~=1.71.0